feat: 支持查询release notes功能
