--- bcs-app/backend/tests/bcs_k8s/releases/conftest.py
@@ -1,128 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) available.
-# Copyright (C) 2017-2019 THL A29 Limited, a Tencent company. All rights reserved.
-# Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://opensource.org/licenses/MIT
-#
-# Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
-# an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
-# specific language governing permissions and limitations under the License.
-#
-import pytest
-
-from backend.resources.cluster.models import CtxCluster
-
-
-@pytest.fixture
-def ctx_cluster(cluster_id, project_id):
-    return CtxCluster.create(id=cluster_id, token="token", project_id=project_id)
-
-
-@pytest.fixture
-def release_name():
-    return "bk-redis"
-
-
-@pytest.fixture
-def revision():
-    return 1
-
-
-@pytest.fixture
-def default_namespace():
-    return "default"
-
-
-@pytest.fixture
-def parsed_release_data():
-    return {
-        "name": "bk-redis",
-        "info": {
-            "first_deployed": "2021-05-21T14:47:53.959134+08:00",
-            "last_deployed": "2021-05-21T14:47:53.959134+08:00",
-            "deleted": "",
-            "description": "Install complete",
-            "status": "deployed",
-            "notes": "1. Get the application URL by running these commands:\n  export POD_NAME=$(kubectl get pods --namespace default -l \"app=bk-redis1,release=bk-redis\" -o jsonpath=\"{.items[0].metadata.name}\")\n  echo \"Visit http://127.0.0.1:8080 to use your application\"\n  kubectl port-forward $POD_NAME 8080:80\n",  # noqa
-        },
-        "chart": {
-            "metadata": {
-                "name": "bk-redis1",
-                "version": "0.1.46",
-                "description": "A Helm chart for Kubernetie",
-                "keywords": ["test"],
-                "icon": "https://raw.githubusercontent.com/grafana/grafana/master/public/img/logo_transparent_400x.png",
-                "apiVersion": "v2",
-                "appVersion": "1.0",
-                "annotations": {"test": "true", "test1": "false"},
-            },
-            "lock": None,
-            "templates": [
-                {
-                    "name": "templates/NOTES.txt",
-                    "data": "MS4gR2V0IHRoZSBhcHBsaWNhdGlvbiBVUkwgYnkgcnVubmluZyB0aGVzZSBjb21tYW5kczoKe3stIGlmIC5WYWx1ZXMuaW5ncmVzcy5lbmFibGVkIH19Cnt7LSByYW5nZSAuVmFsdWVzLmluZ3Jlc3MuaG9zdHMgfX0KICBodHRwe3sgaWYgJC5WYWx1ZXMuaW5ncmVzcy50bHMgfX1ze3sgZW5kIH19Oi8ve3sgLiB9fXt7ICQuVmFsdWVzLmluZ3Jlc3MucGF0aCB9fQp7ey0gZW5kIH19Cnt7LSBlbHNlIGlmIGNvbnRhaW5zICJOb2RlUG9ydCIgLlZhbHVlcy5zZXJ2aWNlLnR5cGUgfX0KICBleHBvcnQgTk9ERV9QT1JUPSQoa3ViZWN0bCBnZXQgLS1uYW1lc3BhY2Uge3sgLlJlbGVhc2UuTmFtZXNwYWNlIH19IC1vIGpzb25wYXRoPSJ7LnNwZWMucG9ydHNbMF0ubm9kZVBvcnR9IiBzZXJ2aWNlcyB7eyB0ZW1wbGF0ZSAicmVkaXMuZnVsbG5hbWUiIC4gfX0pCiAgZXhwb3J0IE5PREVfSVA9JChrdWJlY3RsIGdldCBub2RlcyAtLW5hbWVzcGFjZSB7eyAuUmVsZWFzZS5OYW1lc3BhY2UgfX0gLW8ganNvbnBhdGg9InsuaXRlbXNbMF0uc3RhdHVzLmFkZHJlc3Nlc1swXS5hZGRyZXNzfSIpCiAgZWNobyBodHRwOi8vJE5PREVfSVA6JE5PREVfUE9SVAp7ey0gZWxzZSBpZiBjb250YWlucyAiTG9hZEJhbGFuY2VyIiAuVmFsdWVzLnNlcnZpY2UudHlwZSB9fQogICAgIE5PVEU6IEl0IG1heSB0YWtlIGEgZmV3IG1pbnV0ZXMgZm9yIHRoZSBMb2FkQmFsYW5jZXIgSVAgdG8gYmUgYXZhaWxhYmxlLgogICAgICAgICAgIFlvdSBjYW4gd2F0Y2ggdGhlIHN0YXR1cyBvZiBieSBydW5uaW5nICdrdWJlY3RsIGdldCBzdmMgLXcge3sgdGVtcGxhdGUgInJlZGlzLmZ1bGxuYW1lIiAuIH19JwogIGV4cG9ydCBTRVJWSUNFX0lQPSQoa3ViZWN0bCBnZXQgc3ZjIC0tbmFtZXNwYWNlIHt7IC5SZWxlYXNlLk5hbWVzcGFjZSB9fSB7eyB0ZW1wbGF0ZSAicmVkaXMuZnVsbG5hbWUiIC4gfX0gLW8ganNvbnBhdGg9J3suc3RhdHVzLmxvYWRCYWxhbmNlci5pbmdyZXNzWzBdLmlwfScpCiAgZWNobyBodHRwOi8vJFNFUlZJQ0VfSVA6e3sgLlZhbHVlcy5zZXJ2aWNlLnBvcnQgfX0Ke3stIGVsc2UgaWYgY29udGFpbnMgIkNsdXN0ZXJJUCIgLlZhbHVlcy5zZXJ2aWNlLnR5cGUgfX0KICBleHBvcnQgUE9EX05BTUU9JChrdWJlY3RsIGdldCBwb2RzIC0tbmFtZXNwYWNlIHt7IC5SZWxlYXNlLk5hbWVzcGFjZSB9fSAtbCAiYXBwPXt7IHRlbXBsYXRlICJyZWRpcy5uYW1lIiAuIH19LHJlbGVhc2U9e3sgLlJlbGVhc2UuTmFtZSB9fSIgLW8ganNvbnBhdGg9InsuaXRlbXNbMF0ubWV0YWRhdGEubmFtZX0iKQogIGVjaG8gIlZpc2l0IGh0dHA6Ly8xMjcuMC4wLjE6ODA4MCB0byB1c2UgeW91ciBhcHBsaWNhdGlvbiIKICBrdWJlY3RsIHBvcnQtZm9yd2FyZCAkUE9EX05BTUUgODA4MDo4MAp7ey0gZW5kIH19Cg==",  # noqa
-                },
-                {
-                    "name": "templates/_helpers.tpl",
-                    "data": "e3svKiB2aW06IHNldCBmaWxldHlwZT1tdXN0YWNoZTogKi99fQp7ey8qCkV4cGFuZCB0aGUgbmFtZSBvZiB0aGUgY2hhcnQuCiovfX0Ke3stIGRlZmluZSAicmVkaXMubmFtZSIgLX19Cnt7LSBkZWZhdWx0IC5DaGFydC5OYW1lIC5WYWx1ZXMubmFtZU92ZXJyaWRlIHwgdHJ1bmMgNjMgfCB0cmltU3VmZml4ICItIiAtfX0Ke3stIGVuZCAtfX0KCnt7LyoKQ3JlYXRlIGEgZGVmYXVsdCBmdWxseSBxdWFsaWZpZWQgYXBwIG5hbWUuCldlIHRydW5jYXRlIGF0IDYzIGNoYXJzIGJlY2F1c2Ugc29tZSBLdWJlcm5ldGVzIG5hbWUgZmllbGRzIGFyZSBsaW1pdGVkIHRvIHRoaXMgKGJ5IHRoZSBETlMgbmFtaW5nIHNwZWMpLgpJZiByZWxlYXNlIG5hbWUgY29udGFpbnMgY2hhcnQgbmFtZSBpdCB3aWxsIGJlIHVzZWQgYXMgYSBmdWxsIG5hbWUuCiovfX0Ke3stIGRlZmluZSAicmVkaXMuZnVsbG5hbWUiIC19fQp7ey0gaWYgLlZhbHVlcy5mdWxsbmFtZU92ZXJyaWRlIC19fQp7ey0gLlZhbHVlcy5mdWxsbmFtZU92ZXJyaWRlIHwgdHJ1bmMgNjMgfCB0cmltU3VmZml4ICItIiAtfX0Ke3stIGVsc2UgLX19Cnt7LSAkbmFtZSA6PSBkZWZhdWx0IC5DaGFydC5OYW1lIC5WYWx1ZXMubmFtZU92ZXJyaWRlIC19fQp7ey0gaWYgY29udGFpbnMgJG5hbWUgLlJlbGVhc2UuTmFtZSAtfX0Ke3stIC5SZWxlYXNlLk5hbWUgfCB0cnVuYyA2MyB8IHRyaW1TdWZmaXggIi0iIC19fQp7ey0gZWxzZSAtfX0Ke3stIHByaW50ZiAiJXMtJXMiIC5SZWxlYXNlLk5hbWUgJG5hbWUgfCB0cnVuYyA2MyB8IHRyaW1TdWZmaXggIi0iIC19fQp7ey0gZW5kIC19fQp7ey0gZW5kIC19fQp7ey0gZW5kIC19fQoKe3svKgpDcmVhdGUgY2hhcnQgbmFtZSBhbmQgdmVyc2lvbiBhcyB1c2VkIGJ5IHRoZSBjaGFydCBsYWJlbC4KKi99fQp7ey0gZGVmaW5lICJyZWRpcy5jaGFydCIgLX19Cnt7LSBwcmludGYgIiVzLSVzIiAuQ2hhcnQuTmFtZSAuQ2hhcnQuVmVyc2lvbiB8IHJlcGxhY2UgIisiICJfIiB8IHRydW5jIDYzIHwgdHJpbVN1ZmZpeCAiLSIgLX19Cnt7LSBlbmQgLX19Cg==",  # noqa
-                },
-                {
-                    "name": "templates/deployment.yaml",
-                    "data": "YXBpVmVyc2lvbjogYXBwcy92MQpraW5kOiBEZXBsb3ltZW50Cm1ldGFkYXRhOgogIG5hbWU6IHt7IHRlbXBsYXRlICJyZWRpcy5mdWxsbmFtZSIgLiB9fQogIGxhYmVsczoKICAgIGFwcDoge3sgdGVtcGxhdGUgInJlZGlzLm5hbWUiIC4gfX0KICAgIGNoYXJ0OiB7eyB0ZW1wbGF0ZSAicmVkaXMuY2hhcnQiIC4gfX0KICAgIHJlbGVhc2U6IHt7IC5SZWxlYXNlLk5hbWUgfX0KICAgIGhlcml0YWdlOiB7eyAuUmVsZWFzZS5TZXJ2aWNlIH19CiAgICB0ZXQ0NDogIjMzNDIiCiAgICB0ZXN0MTIzOiB7eyAuVmFsdWVzLnRlc3QudGVzdDEgfX0KICBhbm5vdGF0aW9uczoKICAgIHNpZGVjYXIudGJ1c2FwcC5pby9pbmplY3Q6ICJubyIKICAgIHRlc3QxMjM6ICJ5ZXMiCnNwZWM6CiAgcmVwbGljYXM6IDEKICBzZWxlY3RvcjoKICAgIG1hdGNoTGFiZWxzOgogICAgICBhcHA6IHt7IHRlbXBsYXRlICJyZWRpcy5uYW1lIiAuIH19CiAgICAgIHJlbGVhc2U6IHt7IC5SZWxlYXNlLk5hbWUgfX0KICB0ZW1wbGF0ZToKICAgIG1ldGFkYXRhOgogICAgICBsYWJlbHM6CiAgICAgICAgYXBwOiB7eyB0ZW1wbGF0ZSAicmVkaXMubmFtZSIgLiB9fQogICAgICAgIHJlbGVhc2U6IHt7IC5SZWxlYXNlLk5hbWUgfX0KICAgIHNwZWM6CiAgICAgIGltYWdlUHVsbFNlY3JldHM6CiAgICAgIC0gbmFtZTogInRlc3QxMjMxMjMiCiAgICAgIGNvbnRhaW5lcnM6CiAgICAgICAgLSBuYW1lOiB7eyAuQ2hhcnQuTmFtZSB9fQogICAgICAgICAgaW1hZ2U6ICJ7eyAuVmFsdWVzLmltYWdlLnJlcG9zaXRvcnkgfX06e3sgLlZhbHVlcy5pbWFnZS50YWcgfX0iCiAgICAgICAgICBpbWFnZVB1bGxQb2xpY3k6IHt7IC5WYWx1ZXMuaW1hZ2UucHVsbFBvbGljeSB9fQogICAgICAgICAgZW52OgogICAgICAgICAgLSBuYW1lOiAidGVzdCIKICAgICAgICAgICAgdmFsdWU6ICJ0ZXN0IgogICAgICAgICAgLSBuYW1lOiAidGVzdCIKICAgICAgICAgICAgdmFsdWU6ICJ0ZXN0MTIzIgogICAgICAgICAgcG9ydHM6CiAgICAgICAgICAgIC0gbmFtZTogaHR0cAogICAgICAgICAgICAgIGNvbnRhaW5lclBvcnQ6IDgwCiAgICAgICAgICAgICAgcHJvdG9jb2w6IFRDUAogICAgICAgICAgbGl2ZW5lc3NQcm9iZToKICAgICAgICAgICAgaHR0cEdldDoKICAgICAgICAgICAgICBwYXRoOiAvCiAgICAgICAgICAgICAgcG9ydDogaHR0cAogICAgICAgICAgcmVhZGluZXNzUHJvYmU6CiAgICAgICAgICAgIGh0dHBHZXQ6CiAgICAgICAgICAgICAgcGF0aDogLwogICAgICAgICAgICAgIHBvcnQ6IGh0dHAKICAgICAgICAgIHJlc291cmNlczoKe3sgdG9ZYW1sIC5WYWx1ZXMucmVzb3VyY2VzIHwgaW5kZW50IDEyIH19CiAgICB7ey0gd2l0aCAuVmFsdWVzLm5vZGVTZWxlY3RvciB9fQogICAgICBub2RlU2VsZWN0b3I6Cnt7IHRvWWFtbCAuIHwgaW5kZW50IDggfX0KICAgIHt7LSBlbmQgfX0KICAgIHt7LSB3aXRoIC5WYWx1ZXMuYWZmaW5pdHkgfX0KICAgICAgYWZmaW5pdHk6Cnt7IHRvWWFtbCAuIHwgaW5kZW50IDggfX0KICAgIHt7LSBlbmQgfX0KICAgIHt7LSB3aXRoIC5WYWx1ZXMudG9sZXJhdGlvbnMgfX0KICAgICAgdG9sZXJhdGlvbnM6Cnt7IHRvWWFtbCAuIHwgaW5kZW50IDggfX0KICAgIHt7LSBlbmQgfX0K",  # noqa
-                },
-                {"name": "templates/deployment1.yaml", "data": ""},
-                {
-                    "name": "templates/hpa.yaml",
-                    "data": "YXBpVmVyc2lvbjogYXV0b3NjYWxpbmcvdjJiZXRhMQpraW5kOiBIb3Jpem9udGFsUG9kQXV0b3NjYWxlcgptZXRhZGF0YToKICBuYW1lOiB7eyB0ZW1wbGF0ZSAicmVkaXMuZnVsbG5hbWUiIC4gfX0KICBsYWJlbHM6CiAgICBoZWxtLnNoL2NoYXJ0OiB0ZXN0d2ViLTAuMi42CiAgICBhcHAua3ViZXJuZXRlcy5pby9uYW1lOiB0ZXN0d2ViCiAgICBhcHAua3ViZXJuZXRlcy5pby9pbnN0YW5jZTogdGVzdHdlYi0yMDA1MTIwNzMzCiAgICBhcHAua3ViZXJuZXRlcy5pby92ZXJzaW9uOiB2MS4wLjIuMjAyMDA1MTIKICAgIGFwcC5rdWJlcm5ldGVzLmlvL21hbmFnZWQtYnk6IEhlbG0KICAgIGlvLnRlbmNlbnQucGFhcy5zb3VyY2VfdHlwZTogaGVsbQpzcGVjOgogIHNjYWxlVGFyZ2V0UmVmOgogICAgYXBpVmVyc2lvbjogYXBwcy92MQogICAga2luZDogRGVwbG95bWVudAogICAgbmFtZTogdGVzdHdlYi0yMDA1MTIwNzMzCiAgbWluUmVwbGljYXM6IDEKICBtYXhSZXBsaWNhczogNQogIG1ldHJpY3M6CiAgLSB0eXBlOiBSZXNvdXJjZQogICAgcmVzb3VyY2U6CiAgICAgIG5hbWU6IGNwdQogICAgICB0YXJnZXRBdmVyYWdlVXRpbGl6YXRpb246IDgwCiAgLSB0eXBlOiBSZXNvdXJjZQogICAgcmVzb3VyY2U6CiAgICAgIG5hbWU6IG1lbW9yeQogICAgICB0YXJnZXRBdmVyYWdlVXRpbGl6YXRpb246IDc5Cg==",  # noqa
-                },
-                {
-                    "name": "templates/ingress.yaml",
-                    "data": "e3stIGlmIC5WYWx1ZXMuaW5ncmVzcy5lbmFibGVkIC19fQp7ey0gJGZ1bGxOYW1lIDo9IGluY2x1ZGUgInJlZGlzLmZ1bGxuYW1lIiAuIC19fQp7ey0gJGluZ3Jlc3NQYXRoIDo9IC5WYWx1ZXMuaW5ncmVzcy5wYXRoIC19fQphcGlWZXJzaW9uOiBleHRlbnNpb25zL3YxYmV0YTEKa2luZDogSW5ncmVzcwptZXRhZGF0YToKICBuYW1lOiB7eyAkZnVsbE5hbWUgfX0KICBsYWJlbHM6CiAgICBhcHA6IHt7IHRlbXBsYXRlICJyZWRpcy5uYW1lIiAuIH19CiAgICBjaGFydDoge3sgdGVtcGxhdGUgInJlZGlzLmNoYXJ0IiAuIH19CiAgICByZWxlYXNlOiB7eyAuUmVsZWFzZS5OYW1lIH19CiAgICBoZXJpdGFnZToge3sgLlJlbGVhc2UuU2VydmljZSB9fQp7ey0gd2l0aCAuVmFsdWVzLmluZ3Jlc3MuYW5ub3RhdGlvbnMgfX0KICBhbm5vdGF0aW9uczoKe3sgdG9ZYW1sIC4gfCBpbmRlbnQgNCB9fQp7ey0gZW5kIH19CnNwZWM6Cnt7LSBpZiAuVmFsdWVzLmluZ3Jlc3MudGxzIH19CiAgdGxzOgogIHt7LSByYW5nZSAuVmFsdWVzLmluZ3Jlc3MudGxzIH19CiAgICAtIGhvc3RzOgogICAgICB7ey0gcmFuZ2UgLmhvc3RzIH19CiAgICAgICAgLSB7eyAuIH19CiAgICAgIHt7LSBlbmQgfX0KICAgICAgc2VjcmV0TmFtZToge3sgLnNlY3JldE5hbWUgfX0KICB7ey0gZW5kIH19Cnt7LSBlbmQgfX0KICBydWxlczoKICB7ey0gcmFuZ2UgLlZhbHVlcy5pbmdyZXNzLmhvc3RzIH19CiAgICAtIGhvc3Q6IHt7IC4gfX0KICAgICAgaHR0cDoKICAgICAgICBwYXRoczoKICAgICAgICAgIC0gcGF0aDoge3sgJGluZ3Jlc3NQYXRoIH19CiAgICAgICAgICAgIGJhY2tlbmQ6CiAgICAgICAgICAgICAgc2VydmljZU5hbWU6IHt7ICRmdWxsTmFtZSB9fQogICAgICAgICAgICAgIHNlcnZpY2VQb3J0OiBodHRwCiAge3stIGVuZCB9fQp7ey0gZW5kIH19Cg==",  # noqa
-                },
-                {
-                    "name": "templates/service.yaml",
-                    "data": "YXBpVmVyc2lvbjogdjEKa2luZDogU2VydmljZQptZXRhZGF0YToKICBuYW1lOiB0ZXN0MTIzMjEKICBsYWJlbHM6CiAgICBhcHA6IHt7IHRlbXBsYXRlICJyZWRpcy5uYW1lIiAuIH19CiAgICBjaGFydDoge3sgdGVtcGxhdGUgInJlZGlzLmNoYXJ0IiAuIH19CiAgICByZWxlYXNlOiB7eyAuUmVsZWFzZS5OYW1lIH19CiAgICBoZXJpdGFnZToge3sgLlJlbGVhc2UuU2VydmljZSB9fQpzcGVjOgogIHR5cGU6IHt7IC5WYWx1ZXMuc2VydmljZS50eXBlIH19CiAgcG9ydHM6CiAgICAtIHBvcnQ6IHt7IC5WYWx1ZXMuc2VydmljZS5wb3J0IH19CiAgICAgIHRhcmdldFBvcnQ6IGh0dHAKICAgICAgcHJvdG9jb2w6IFRDUAogICAgICBuYW1lOiBodHRwCiAgc2VsZWN0b3I6CiAgICBhcHA6IHt7IHRlbXBsYXRlICJyZWRpcy5uYW1lIiAuIH19CiAgICByZWxlYXNlOiB7eyAuUmVsZWFzZS5OYW1lIH19Cg==",  # noqa
-                },
-            ],
-            "values": {
-                "affinity": {},
-                "image": {"pullPolicy": "IfNotPresent", "repository": "demo.io/nginx", "tag": "latest"},
-                "ingress": {
-                    "annotations": {},
-                    "enabled": False,
-                    "hosts": ["chart-example.local"],
-                    "path": "/",
-                    "tls": [],
-                },
-                "labels": [{"app-name": "test-db"}],
-                "nodeSelector": {},
-                "replicaCount": 1,
-                "resources": {"limits": {"cpu": "100m", "memory": "128Mi"}},
-                "service": {"port": 8080, "type": "ClusterIP"},
-                "test": {"test1": "test1"},
-                "tolerations": [],
-            },
-            "schema": None,
-            "files": [
-                {
-                    "name": ".helmignore",
-                    "data": "IyBQYXR0ZXJucyB0byBpZ25vcmUgd2hlbiBidWlsZGluZyBwYWNrYWdlcy4KIyBUaGlzIHN1cHBvcnRzIHNoZWxsIGdsb2IgbWF0Y2hpbmcsIHJlbGF0aXZlIHBhdGggbWF0Y2hpbmcsIGFuZAojIG5lZ2F0aW9uIChwcmVmaXhlZCB3aXRoICEpLiBPbmx5IG9uZSBwYXR0ZXJuIHBlciBsaW5lLgouRFNfU3RvcmUKIyBDb21tb24gVkNTIGRpcnMKLmdpdC8KLmdpdGlnbm9yZQouYnpyLwouYnpyaWdub3JlCi5oZy8KLmhnaWdub3JlCi5zdm4vCiMgQ29tbW9uIGJhY2t1cCBmaWxlcwoqLnN3cAoqLmJhawoqLnRtcAoqfgojIFZhcmlvdXMgSURFcwoucHJvamVjdAouaWRlYS8KKi50bXByb2oK",  # noqa
-                },
-                {
-                    "name": "bcs-values/test.yaml",
-                    "data": "IyBEZWZhdWx0IHZhbHVlcyBmb3IgcmVkaXMuCiMgVGhpcyBpcyBhIFlBTUwtZm9ybWF0dGVkIGZpbGUuCiMgRGVjbGFyZSB2YXJpYWJsZXMgdG8gYmUgcGFzc2VkIGludG8geW91ciB0ZW1wbGF0ZXMuCgpyZXBsaWNhQ291bnQ6IDEKCmltYWdlOgogIHJlcG9zaXRvcnk6ICJ0ZXN0LmFydGlmYWN0b3J5Lm9hLmNvbTo4NDQzL3BhYXMvazhzdGVzdC9uZ2lueCIKICB0YWc6IGxhdGVzdAogIHB1bGxQb2xpY3k6IElmTm90UHJlc2VudAoKc2VydmljZToKICB0eXBlOiBDbHVzdGVySVAKICBwb3J0OiA4MDgwCgppbmdyZXNzOgogIGVuYWJsZWQ6IGZhbHNlCiAgYW5ub3RhdGlvbnM6IHt9CiAgICAjIGt1YmVybmV0ZXMuaW8vaW5ncmVzcy5jbGFzczogbmdpbngKICAgICMga3ViZXJuZXRlcy5pby90bHMtYWNtZTogInRydWUiCiAgcGF0aDogLwogIGhvc3RzOgogICAgLSBjaGFydC1leGFtcGxlLmxvY2FsCiAgdGxzOiBbXQogICMgIC0gc2VjcmV0TmFtZTogY2hhcnQtZXhhbXBsZS10bHMKICAjICAgIGhvc3RzOgogICMgICAgICAtIGNoYXJ0LWV4YW1wbGUubG9jYWwKCnJlc291cmNlczoKICAjIFdlIHVzdWFsbHkgcmVjb21tZW5kIG5vdCB0byBzcGVjaWZ5IGRlZmF1bHQgcmVzb3VyY2VzIGFuZCB0byBsZWF2ZSB0aGlzIGFzIGEgY29uc2Npb3VzCiAgIyBjaG9pY2UgZm9yIHRoZSB1c2VyLiBUaGlzIGFsc28gaW5jcmVhc2VzIGNoYW5jZXMgY2hhcnRzIHJ1biBvbiBlbnZpcm9ubWVudHMgd2l0aCBsaXR0bGUKICAjIHJlc291cmNlcywgc3VjaCBhcyBNaW5pa3ViZS4gSWYgeW91IGRvIHdhbnQgdG8gc3BlY2lmeSByZXNvdXJjZXMsIHVuY29tbWVudCB0aGUgZm9sbG93aW5nCiAgIyBsaW5lcywgYWRqdXN0IHRoZW0gYXMgbmVjZXNzYXJ5LCBhbmQgcmVtb3ZlIHRoZSBjdXJseSBicmFjZXMgYWZ0ZXIgJ3Jlc291cmNlczonLgogIGxpbWl0czoKICAgIGNwdTogMTAwbQogICAgbWVtb3J5OiAxMjhNaQogICMgcmVxdWVzdHM6CiAgIyAgY3B1OiAxMDBtCiAgIyAgbWVtb3J5OiAxMjhNaQoKbm9kZVNlbGVjdG9yOiB7fQoKdG9sZXJhdGlvbnM6IFtdCgphZmZpbml0eToge30KCmxhYmVsczoKICAtIGFwcC1uYW1lOiB0ZXN0LWRiCgp0ZXN0OgogIHRlc3QxOiAidGVzdDE2NjY3IiA=",  # noqa
-                },
-            ],
-        },
-        "manifest": "---\n# Source: bk-redis1/templates/service.yaml\napiVersion: v1\nkind: Service\nmetadata:\n  name: test12321\n  labels:\n    app: bk-redis1\n    chart: bk-redis1-0.1.46\n    release: bk-redis\n    heritage: Helm\nspec:\n  type: ClusterIP\n  ports:\n    - port: 8080\n      targetPort: http\n      protocol: TCP\n      name: http\n  selector:\n    app: bk-redis1\n    release: bk-redis\n---\n# Source: bk-redis1/templates/deployment.yaml\napiVersion: apps/v1\nkind: Deployment\nmetadata:\n  name: bk-redis-bk-redis1\n  labels:\n    app: bk-redis1\n    chart: bk-redis1-0.1.46\n    release: bk-redis\n    heritage: Helm\n    tet44: \"3342\"\n    test123: test1\n  annotations:\n    sidecar.tbusapp.io/inject: \"no\"\n    test123: \"yes\"\nspec:\n  replicas: 1\n  selector:\n    matchLabels:\n      app: bk-redis1\n      release: bk-redis\n  template:\n    metadata:\n      labels:\n        app: bk-redis1\n        release: bk-redis\n    spec:\n      imagePullSecrets:\n      - name: \"test123123\"\n      containers:\n        - name: bk-redis1\n          image: \"demo.io/nginx:latest\"\n          imagePullPolicy: IfNotPresent\n          env:\n          - name: \"test\"\n            value: \"test\"\n          - name: \"test\"\n            value: \"test123\"\n          ports:\n            - name: http\n              containerPort: 80\n              protocol: TCP\n          livenessProbe:\n            httpGet:\n              path: /\n              port: http\n          readinessProbe:\n            httpGet:\n              path: /\n              port: http\n          resources:\n            limits:\n              cpu: 100m\n              memory: 128Mi\n---\n# Source: bk-redis1/templates/hpa.yaml\napiVersion: autoscaling/v2beta1\nkind: HorizontalPodAutoscaler\nmetadata:\n  name: bk-redis-bk-redis1\n  labels:\n    helm.sh/chart: testweb-0.2.6\n    app.kubernetes.io/name: testweb\n    app.kubernetes.io/instance: testweb-2005120733\n    app.kubernetes.io/version: v1.0.2.20200512\n    app.kubernetes.io/managed-by: Helm\n    io.tencent.paas.source_type: helm\nspec:\n  scaleTargetRef:\n    apiVersion: apps/v1\n    kind: Deployment\n    name: testweb-2005120733\n  minReplicas: 1\n  maxReplicas: 5\n  metrics:\n  - type: Resource\n    resource:\n      name: cpu\n      targetAverageUtilization: 80\n  - type: Resource\n    resource:\n      name: memory\n      targetAverageUtilization: 79\n",  # noqa
-        "version": 1,
-        "namespace": "default",
-    }
-
-
-@pytest.fixture
-def release_secret_data():
-    return """SDRzSUFBQUFBQUFDLzl4NzIzT2pPTGZ2ditMaWZBL24xTWtGc0VuSHJwb0hRd3pHc2VuNHh1M0wxQlFTQkxBbFlBell4bFA5disrU3VCZzc3dTdwcWYzdGg5MVZYUWxDTEszNytpMUorWXVKSE93eEF3WnM3M2VlRzZiTUhSTkdIekV6K0l2NUNIZHA5b2ZySlNndVBKY1pNRHpMYy9lc2NNOXpLNjQzNkgwWkNOMkh2dERudXIzL3p6NFBXSmE1WTVEenk1KzRIdkl5T3BrK3BIQVhKbGtZUjh5QVVhTTBjeERxd0JnblpCSnp4NlNaaytVcE0yQ2FSZTZZS000OE1zUTlkQlF2NjJTQjEzR1NCSVhRSVhRNjY4VzBBNHJPTG8raU1QTEo2OVFqSkxFVHVlbmdQZXAwdkdNUzc3TE8yOWVYUDdUaGJQVGJ2Lzd2TmdjZXpGREg5N0pPRXJ0cDUvNmVLQ3BOSE9oMVhPL0R5VkhXdVVlZGQ4WkprdDlxNVhGM093OTVUdW8xSSs5TTV6N3ViTkk0U3B3cytPMmQrZXNoekR5Yy9wdjkvUUY3bWVNNm1mTkFLSDk3Wi80ZlpRVUdjZWVkMGNNMHpEcEJsaVdEeDBlTy8vTEFQckFQM09DWmZXWTdXZHpKVTY5VHhQbXVMZWM3UTc2dkdTY0MzWC9FdTRPemN6di9xaVhyRUFLRFovWTlZcjdkTVRCd2RobXhkTTBKK2YzS0h6am1qdGw3dTdRMENQdkFQZlNlUHRscDJCbDdDSGNvdmM1SHZPdTg1c0RiUlY0V0VwTnR2ZUlRNzl5VUdmeWJ5YncwWTM2L1kwSklQeVFDcG9QSHg1MXplUERETE1oQm5ubzdHRWVaRjJVUE1NYVAvczc1Y0NLbitZbWROUE4yajBrT1VBZ2ZRK3cvb3RpUC84aDJUcFFtenM2THNqOTZMSHQ4U0NLZnVXT2NKTlFiNXZjOEhVbk9JOXdEY1VBbml1S01xakFsQ3FBc0RwaHNseFBteVJQSERKZ1BCNlVlOCszYkhZTml1R1VHVVk0UWVZc1Q1RkR2KzNlanVtYndVZnU2R2kwZnNtTkdORWIxeTh5V1BYL0I2Nnc2WHNUMlVnemdXRXdkUXd0Y0JlMUJLT3JyN2NHM29xMFBJejBIR09WMkliS09vcC9zcGJnQlBKZFpockNGcC9qVjY2YVpxaUNzU29KaEdVZk9ObWU1WXdnUnhQb0pGZ0lDV0E2Qm9tL1ZNZGVYb3V6TGRDa1dsaUZFOW5LWTYxaE9YVU0vVFFuOTdnVEI3aXgzbFA3SkhjLzhENU45VlNVeGRzZUxnOWROZmNldy9NbnROVmhBNTNNbk1zODJCTHJXMS9CNVQ1Nm5vZGovTUxNdnFqUy91UjVVWk5hUnhQN0hQUG5pRld6emZjVXJBbU1OVWZrVWJRK2lSZUFZd2ttVkpsOEJ2MEJycFYrNGt1cFBrUjJBc1k1Z0laeHNjOEk3aG9hbTBVS0F5cnFXQTNsamNRK2p1Yi9hOWtjTHZUOWZjWlAxMjNJZU8xMDl0QTJOQlpJWTJlYmNueTY1M0RJNEJMdGlZUEZybjhxQUpnZ29lZ0Q1ZGI3Q2NtYWIyc0V5TkVUNFZDVnVyeXJKQ2ZEQ3dUSVg4ZHR5OG1VYWFRZmJJTEwxQzNlc2dabk01Z0QzdDdaT2VGajAxVkJzK0lTRitNVXJSTlkydUFOUVpOWmVEa09JOWExanpuSTcwbE9nQ0FFdzFxRXE5WWdzaVJRT2Zkc01EcUE3WWRXUjhMWVk2UjlMZmRpZlNNSE9OU2JJNmk1U1ZYR1JLNGs1MFJFc2h0blVJRFQwRTFUa2piMGs2dzN6TmRaVDI1QlA5bEw0MnBiM3cyVDlxZkhzT3hIUnR4aTRpdDlYb3pSM3pBVUNaaWtMN0M0Q2QwenNLRy90TWJHamhpQ1hIc3lsRU5qS29yQk43ZlN4VkV0ZURTMEdSZWxIeENjbVo1NmY2dC9Yby81U0g5YjJQeElmVCt5UStMbkFXZ2JLWVRFTVYwby9zRWVUQUNoeWJ2RjZvWVl0LzQwMEJDTTdzZmgxN283UndWNFNmNHA5VlJyNlJFZjZhUDJramhDcktsemdMVVhXTWpLa0tpUGZ4bnBYVmJnRVJEcHJtelBmeHYyaWlzY1o0T1h0SE11cFpRZ2IyMVQ5cFQ3MFhlWFp0L0RhdDB3N2NJeGpZT0VqbXZyVk92Vi9HZTNkcGJpeGpKN3Y4akpyOGI3dktnRlN4eHBybVFzT0Z1TGVEc1hRVzRxRmF3ZzBqbFRKdmJiZHljVXpmMnBDNm4rdW9tZFFPUWF1c3ZiVmFJSnNCWjJtMk9hQWNxUytTblJCZkhGeWlIMVYwWHZVN3lSeHRkQW54bkt0eVNhTDVyZDhIWGJ0alNxeEdiandhUktyd3RJMmpzZ3lOVFRkWHZoTy8yUDVTLzc2eVpjbTNiVGxQOGU5WlN3a3l6Z0dBR3NJaGtJQ3NFdjl4emlKN2hTanc4Y1MzdllqV1pQWHlKN00yZEtYeWppOWxRZkt1Q2Q1b015VmVncjVOYzFuRnQvUFhVVk9RRFR6MWEyV3VxYkcydVprc3Y3Rm5MSWU5VWNtSzRpcjlmcFdIQjRBdnpqOXVwNkhHWkNHb1dXS2h6ZVNQOGNrL3NUVU1oZElsU2FGYlN3U1dBZ1g5cCtPbXp6VnY1bTNLRjMxcC9FTkRKMjFqRVhnS3FPODVKa05YK2ZVdHphTzh1eXJ5RTRnVCtJcFlOM3g4R2xhUEI5bkc1alBwTjVodWhrOWZYMFo5bWFTeUlKQzVJaXVQYVBQd2ZDNnhxbEVoMmRkbGJyTVNBeTZ2RnpZMG5EYjBxdFBhYjdFdmRud3FrNzR2LzNHZkx1N1VYVC9DRHlVZUx2MElVdlF1ZTU2M1hUL0dvcThZN0JQNmxnajlzR09jVVEwYjZ5NGpQaUFaV2l4dllyOTE3QmYxYVhuUDZVdGlTczV0eVZTaDljK0tQVkpZcGsrVzN3UXdHaWVTMkc4UC92YUF0bWsxclZpcFB4TzlhZG1VK08ydG1FSHJuRmtWVWw0Y1JTNWNLVXlKN2RyT3YxdTNlZHRjMUk0eGdLcDQ0UHZqaWNjd0ROZjI4ejhEMGxrSVViWnVxdGpHNk9lS3FtWkdnNnpsdC9udGxRKzAzV0wrSFhlblNEcVR5UVhLanEyVEQwbCtuQ05ZK290eGFOcnlLbGoySWx0ekgzaWgyb1oyN21FWEtTT0Z5Ui9iY3J2WlZaOXNVNnFvc1dXT1RtcHlnUlp2RXh0RC9rKzBkT1UyQmxpQWJtS2Zxcm8rRFpHQ0NpTGs2ckloYjBrdnNFbExzVXFpejNKdzQ0NTgxK1ZpVkRsNU5FS3phamVhYzRjMHhxYlRQMWtZb2RpVWNkUlRic2QyNVZ0YXBzbHJpUjJIZU9ZRWo3VnNYNHE1WnY1MXJLVXZaSHpKN2E4ekhkY2cyRklibW5sRDByemsvMWE4Mzg2OTlkdFRYUGMyY2VHMjFMMjRkUGJQL1czSy9uYStwMVVPditjYjFvOGZjcDE2MUtPU00rdFlzalBDdkdaK0pSamNDdlhzTEZqK3I0YXNoZDZMYkZCaStaWUxCeERZTzF3R0U3TVdUWXhaK0d0ZFdyK2ZuMDlZZnYzbmlrTzM3LzZ5UXZFT3EzVFZ6NFhBRHozWGF3WGtLZjRQb0EwTitwYjllemZtOUlXWW1vWkV3U2szdXM1LzdBMFBoMURhT2YrYXY1RkxqbEFqSEpYc1h3MTFFL1RwWDRpdFdGZTVhYktKczJ6ZnVibldTVTRUamxTREtpR2FhaEtrdzgxTEhWRTRwekdkK21IQ2RBMXpzWjI0a25EY0hxWnl4Q1JrejUvTnkrWERUc21qVjNoNEZacXRrd3hhVmpheERUbHdLTFB6K2JKempHRTdkZFFITm1tbUlJdXlteERZQ1hNSVZlUnQ1YTVDTDRTR0ZhYStZbVcxdHZsOGh4ZWhPMndnb2xFYkt5bnBKMmlFRTZSRC9BbC9nSDB1b0E0MVRjMDliRmZ3Ky9EbzhvbExyODdsK3luMjVCZzNWb2pRQkFqMWpKY1ZLNXpBZU5YTlV5aFpURWtjRlJrYlhQT2FpK3hyMjVtSisxRkRjL2pHanRicWFlYXpobEtMeERzem5PU3B0MlhVUTExQW9DRnZVdGFOYU9mTjNvYWE0bXQ2QnZMVkhOWG1YQ1FsdzlRRWhKUTlCT0FFMlIxNTArcU5NbEJvVmJ6S2UzamJETWo0NEp0emtLcGJKV2VDRjhRNndlZ29JMWx6cDdVbHhGWjkwUjEwVjNzNGFhMkRSZTRpaGF2RkRrazZlQnJBNzhKdkJqK3lQWVhVS25VdzYvb3YyWFRWY1BMcGYrVmZKVGhPeTVscXRzQzRzcy84bzFQUGluOUtuL1VIbzB1Uy81UVJueGxQZFpUSUd2STZrNlFlOEVYVzZhblZleXJVV01iOGo4ODAyamFiZ1NqUzVtbVM1SHF0UGFoeXpSekpZYzA5QjJEQzJ3aWh6UzU5RGxjOGptTlNBcnFueHh6c1lmUmxzaDFEZXNUWU1pUnZTUnRJU1R2d3pZL3FpU1c3M1dSdEVaendCOFRxN3V0OWRiYXVxQjg1SkRxUmR3VG4vTnU4R3NiQXYvMXFyMDd5endNYVl4SWxXKzMvcnRVTGlvbmpUUDF2NEVHaWRWck91WDJ3cVZOcnUzcWpCY3NIRjYxcU5kMlJSUjZQNmt2L3VFenJhRVB4NU85cS9RM2dEODhxZkxpWlgxRkR5aUl0d21kcmphSHVCODI4ZEcyUGVGajVDTDM1Zk03VlJMcDFzblhjTGkvdVQ2UjgrVzJMS1RjMmdyS1NjdTRIay8yRmw1LzFnZHRVOFN4YmM2ZmJ0T1hXZWNsOXFlSEczb2FWN29wVzUxTDNrbko1UHNjSksxcnVSWG51MHJmdGd3dWJmc2J4UG9KZFBYQzRuVmFRa2t0SS9WTGZSa1ZyVnhOeTd6TEk5YVIyckVoN0cxRlh6VjU4Q0kvbEZzOGE1N1VBSTBGWGZXSmxHRUNudzFESnUxamZybWUzOG9WNTNKOU5kWjF6RVhjNXQ4aThNZ1FFbmU4UGM4bE9lMDgvaDlaMTFYNnFXMU95bll4bWwyc2ZmWHVINi8vTTRqQ1hXR1UyL09EeFBrcGx0RlowTlUybG5GTUFJWjdkek1KYlhNUnRMQ05DcnFUeE1NVVZxZHJwYitkbjc5QjBFOHlNdDlXWk5haThkWE92WDk3TytiMWMzMFNZOXM0WnROSWk2ZDhnMkZvem5GNVBaeXVodmtzN1BIVlhGSmpjN3FOWkU1eTIxeVVPYm5vMTd3MDMvMWtmZ0lpMG1ZTEc1S2phQTRjdThnSzJXTDJNdVJtSy9XZ25XYW5uOUFncmNtSjRKR3ZvY2pQbHIzRGRLUG1zODJ3cG5IR2NwS3dhN2VmVTR6MlU1NExBSllqMjVoblZyUjlVa2NCQWtxTnM5Q2U0Q0NBTlFTaWVRNFZPWUNGVU1md1I3bFZFUHVPb3FkZ25weWdvbTlvblJpWHR0SkpPOHZyN0JycnVLNGYzOGUyNVh1SFI3bjlFdnNMaFdDZ3ZnQU1QWGVyWEZmbjhoL3BDUmlJWU1Gci9KUlpackFrbU5reHRBQ2VZbCtqZUpkREJNWmIzZElIcGt1UjlVeVIyRzlwbTlyZU5TY2J1K0xybkx0YWViWEcyWXAyY00rNWlMWE1TV1NiQzVHME9hU2U2K1lpQVFwNnNzaFB2dGZVbDMrK0hvZUEwUys4WDFvVEN0OXZSY0xJMzNscGVoVzdmL3RJcGRVS1RoUzZIVnUyMFM5eFgxVlFidkZIenY3Umx1M0Y5L1dSaURZbnRaRFN1TDArclpYVnR3RlVrTkdLQStRUjdCdHBDZUNGMDdSckhTMnNzOVpxOUZyNzE3S21jL2hSUGhsdWFlNFlYZURmNjd6eEQvQjIxZUwrc0xjcWM5RFZkODMyem8yK3A5VDVlVzVzbTVQRVZlU0l4TXoxTnVpYTF3c1hvM0tMdGRMOTU1cDdQcDZ5RENFSDNjVlZEZnJjRDEzVy9wNy9RYkFvSnJhWSs5ck40NjBLcTVkdGMyS0h0OWQzbGVPcGxvMzhYdWFabngvanRiOVRwV0dtS3NFZWRoZnRmb255QTdHYzIvemFuK0x5L1VWdlZHSlZxdS9MbnVsR0RTZXh5K3NiaUhWMlZlVXJxdnVvNmpzdWZlbjJVVjlEVHl4Y1V2Tm9uM25GWjdzWHFJNEtidkJleVR1dnNIL3ZnaytLSXk5d2FJay9tNzYyaGFOcmJFaGsrUlNqVjdvcWM5VWtzUGlNeUhJYmE5Yit0Mjd0VlVnTHVpL3h2YjZwNnUycUl5NTlEcnEwUnRPakVMTEdlWHYzaHA5OU4vZWwzbTRmUXU4bnVNWGRuSE5IRXp2ejcrYU9wbGVaYlViL1czTkdxOTdUNDZCUC9XVmo0NlZBNjF5OXpsVy9salc5eGZlL0w0OTVMMkp2RVVEc0lsZSszWmQ4djFkcmJOVDREV3ozRGYvY1BuOVB6OVFQZjc5ajlnN0tQWHEvd3ZuNENLTXdLNWpCWDkvdW1CQTd2a2VHa3h5aHR4aUZzR0FHalBxaHhkbmJ6a3U5S0dQdW1KMlh4R21ZeGJ1Q1h2ekI4VU1ZUDBaK0dCMlpPeVp6ZkdiQVVPZk9HRUt4TE81MHFZdWJIZC91R0M5eUFQSmNaa0N2Yzl3eFFaeG05RjRLdmJweTd4MGRuQ0R2QWNYUVFjenZkMHppWkFFellCN0pNb2hNL1AzYkhZTWM0S0h5eG9lVEpQZE5oS1hadlF1b3NGSHNla3NQZVRDTGQrWENPNC9lMDVIaVBNcVlBVWNHMGpqZndWSWxLTVJoUm4rRFNjNE1HSTVsTVhQSFlBK1hJblA4OHl5a2wwK3E2S1VLaTNjWlEyOEUzVEZaa1JBZUpKU25tYmRUMzRnZXluc3NmelUzV01xZjVFV012RjJ0RlNwUkNnTVBPL1dsbG84UVhWNW9lUWc4aEVNL2luZmVPV09vaFVpeUlVdlFPU3pvb1Y5aTg4SWU0clh2OGdFQ29SaTZCa3BwbDE2SUI4dlFkZ1Nwd2FMM3FoYmkybEhRU1IxckhLVFJRTEs0UmpxU1ZGWGNGUENxRHd5WnRmaUFkRTFwdVJzbXM0NXBJM1ZNRHk4djN5dHliZy9qamFvSXlPYkw2cXhLd1FGaUhUdG1nR3lwN2pKSHlUUVUzd0ErQ3FyU3orMGxyUUJVQm5Vc0loZ1N4Q3lncVIvbkMxbjdXSlBPRzY4SnZ5K0E1ekxBOTN4OXE2MVVaWkhBYVBZNnhXN2lTcy9sVHdWRkFQY0xleDduVnBRVTAwUDUwekhjSEhRblNBcUYyQzdJM0NCcWpaMWMzTnRMNGN5ZjgvME1FTDdMU3NMQjZvQVNIdUkvcDVIV2hjUDR6eW1lQkE1OVhtVGsrY09QTjZwc0J4Q2p2V3ZPL09WNkljTkRuSk5zNEdCOTR3N2ozREVXeUZvK3Y3NkdBZ3RNc1FCOGZOSDlBcGplbDhINVNEemtxaTZvaFRocWpvM0dkUTBXTWVpcWZ0MXZFdTUxSlVoZ0laTC9nU29qY2JVKzBDTmRZaVY2dEtmWUNWRFdkTzVDMFRlZ1BQcmpMWE9TV01Za3RjMVpjODBDS3ZLcFBDSkJPUm1yanBIUHZTNVowMCtLdXIrWjgzME8wSDJzMGF0VTdTbVdlZnBpWDdIWlU1dGl1WEFWaEMyYUF5ZkNGUGVES2RiMllCWDN0SmY1YWRvVkE4dWM3WjFUY0NyMzZmcTV6YVBjby90MUltc1o4RW1sdFVnLzBWNXRmTFh2T0VKNGhmdnNtdTRUMFg3dXRjbnRaZDJzKzZBWE1OYkpHc1ZTSjdsY1BKUTFmdGlia1o3SlR4cXNVNTV6NkRuVmxVSHlQN0dGaGtndXZzS3FKSS9YZFdPaktobG5ZYjBBbUY1MXlSM2plZC9xS1lnZFRxUS9CTmhOUU9SWDlXVG0zK2k3V1RDZVpaYWgxZnZHaFd1c3c3Syt0ZmJScnZEbXREbmU0cENueUtTR0kzcjlnNWZUQnR1R0lqQXA5cG1WMk9zS1QxYm5KNWx0QmdHcFRmYVNJN3k4VXZuS2ZiN1dtck1hTzJYMTJjelUwSHNXOVoxMURwVCt4aklPcjFKMHVZZEhhY2t1UFJaMkRUa0Y0eTN4YjNybGptSXFSZGk3NWRVR2lnTWN3eGJLbzJHWkErTzVmN0hYVjEwWEFJVkk2aUZ2TDBXV1pqdEZQcW5LaUI2aFFsNUxRRmN2OXpvS29xTitZdkhyOWxVa2p2ak1OS3d5cFNLbmtILzJIVU1ndWdrZ1hVZUx5MnRLOVZFM3lhSVREb1RpSG9RaUFwR2RRTnpQNlo3Q2VGYjFPMkxxbUFzV0tPdFM1dlplWm5Id1lWZmZPQkk5cHRRY1EwaW9IeXg3L3RLd2FCeXF5bUt2anQyQTlEY2tObUZYUkJhUHNMY1VpNmFuTjJlcE90WnppMlkwUFhlcnF4TTI3cWRBNlhlSi8xVnkwMnhMMXJXTXhaK3VxWlhYSWcyV0hzc0RyRytJNzF2bVJKaEs1VkVxeEhvR3VxUUtsRWVucmpsSnZhVVlRaXlYZWpCczFqWlZmOUs5c0c5RXIyc3B4d1FZaUQyZjkya0hkeFg3czlYd0FDcnNEUXhDZnlKOERZZkgyU2JRbk1vdklkYVBwTmRxY0Z3eDlLMnV5TkY1TDJKV2o5MzQvclc4QXFnUlhMbHhsWDVCOE5MSFBINjkzc3RVNWN5Vi9DU3dzWjBBakZpUFlsTDJWY0lYNTVSWnViZkZYV0R2cWJFSUpUK2h2MWNZbFo3ck5HY05MeU5lMjFoZE5Sd1NQUGJ0anNGT0ZINlVWMTN2NysvZm8vL1RXVkpBTXVnMFYzOGZiM2NONzlINVd1MmdzK2ZlbzIwWXVZUE9zcHp6SHRYWGllbmRhbEptQmgwS1B2Z3V6NUdoRWtEUnQ1Mk9reVN0SmNzeGlzWmFvL2ZsdmVQeVpYWEYrdnk2SEE2OFhaZzV2amVnZDVIZm96VHhJRjJDWUtOQnA0RkdaSWhBcDNyOWUvbzBvRGVqeTVGT0ozTjJ2cGU5MGZFZ3k1SjZQTm5GV1F4ak5PaXNwTGQ2c0JTd25wWldzTzhIMHQwUTRHOVk0T3I0L05JSVRwS2tqMmRMdkRSemJ4cWpKbjkvd2RqL2lGV29kcjJzMXh0MDNwbHV0OGVYdDlZN3RZTlVua0xHV3VDOVlpb05YUTg2dTRjTTVLbVRKS1FKQ0tPTkJ6TkNMSW8va1hwbkNpOGxvNDB2VkRBOEhYUzRHOGJDVGdhRGFWc050eFh4SFdsclc5WFUyb29uLzlBbDVlL1IvcTR1R3luSVA5bzN2ZVVJTFQyNDg3SXoyZnZLeU85TXBRZU83OWFhNlhSZ0hHVk9HSG03TmgvM1YzN1JZcVZhaUpDNzZMc0daYjkxSm56QlU5bkhEVHJ0TnE0OTBZdjJnL2J6SmMrWFJEc2RpazV2di95RkR5L1VRSU81bFFTdTZiV0R2djdYcU82dHloZlhFMjVtQjJyNGNPOUZYcHErN1dMZ1hhMUlWbEs4YlBDSm1KTUZnODdqcCtGUE9hbDBHTWNOLzhNclZPM3FGWm15Yy8xRUd5YjVvRU82Mk9zWFpVczc2TkNPOW0vbHZmcU03U3JoNVZtY1FnZUZrZis0NTRHWE9VM3lHOGU3OEVSc2hkNWlkMWpOODNiL05CT1MzdmNoRFI2cjdFZGM2ZUNCZS9hQmYzaHFjdVhEdHZxREZ5K2xJZExVdklNSHZqY3BqTkxNaWVCNTRqM1BzZ0xIczErNjNlOTlzejlYM1FmQ0FjK1czM3h2UG5ZaXgvZmNlMUMwRTNBWVAyUmVCRWtwU1J3bmZTaE4rMGRaS0lQTDZrbVZ0NkwxY09GOU5OWGhSdWtoTHo2WG4zT0Z2QzBsRHFQRlpWTEd6dkU4SXRBUkw5dUZzRFRJZlZYUEY1VkQxdFduZkJwY0ZtV1k1SmNsZmJqM2RvN3ZyYk1RaFNkYVcrcEEvaVc2cFJmL25QU1gvbnZVL3ZzcDdvNXAvb3FNN21UUnZ5Tmp2djFYQUFBQS8vOTc3bWRuaGpjQUFBPT0="""  # noqa
