+++ bcs-app/backend/tests/helm/releases/conftest.py
@@ -11,14 +11,38 @@
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 #
-from rest_framework.response import Response
+import pytest
 
-from backend.bcs_web.viewsets import SystemViewSet
+from backend.resources.cluster.models import CtxCluster
 
-from .utils import release_secret
+from .contents import release_data
 
 
-class HelmReleaseViewSet(SystemViewSet):
-    def get_notes(self, request, project_id, cluster_id, namespace, release_name):
-        """查询release下的note"""
-        return Response(release_secret.get_release_notes(request.ctx_cluster, namespace, release_name))
+@pytest.fixture
+def ctx_cluster(cluster_id, project_id):
+    return CtxCluster.create(id=cluster_id, token="token", project_id=project_id)
+
+
+@pytest.fixture
+def release_name():
+    return "bk-redis"
+
+
+@pytest.fixture
+def revision():
+    return 1
+
+
+@pytest.fixture
+def default_namespace():
+    return "default"
+
+
+@pytest.fixture
+def parsed_release_data():
+    return release_data.release_data
+
+
+@pytest.fixture
+def release_secret_data():
+    return release_data.secret_data
